Add programmes index page
This also adds factories for programmes and programme prerequisites,
and adds a missing export in `services/index.ts`.

We've taken the approach to typing RequestHandlers from the CAS team as
they provide a slightly more explicit response type than `Promise<void>`
which we've specified on other projects. We're not 100% sure of the
benefit, but it felt right to join up with their approach.

We may later rename the key/value fields on the ProgrammePrerequisite
type to make these more domain-specific, but this will require an API
change.

Co-authored-by: George <george@dxw.com>