Refactor routes and replace index with dashboard
Apologies, this is a large commit but adding the new DashboardController
in the way we want to write the code required some refactoring of how we
structure our service. I've tried to keep this as simple as possible.

This commit adds a new DashboardController which is responsible for
rendering the page users see when they log into the service. The "Find
an accredited programme" link doesn't currently go anywhere, but will
point to the new "Find" journey, which we'll be building very soon.

Controllers will be passed to the `routes()` function directly to allow
us to use the controller actions directly when mapping them to a
specific route. This has required some re-jigging further up the chain,
via the new `controllers/index.ts` file that will serve as a "hub" of
sorts for all the controllers in the application that can then be
injected into the running application.

We no longer need to test the index route here, as we're testing it in
the Dashboard Controller test in more of a unit-focused way, rather than
making assertions on the http request/response body - we'll cover that
behaviour in Cypress tests.