Flesh out programme list
For now, the tags are hard-coded because the API doesn't provide that
information, and the summary list capitalisation needs fixing, but that
will likely be an API responsibility

The `parseHtml` method in `integration_tests/pages/page.ts` is taken
from the Approved Premises repo[1]

We add Lodash here for its `omit` method, which is used to remove keys
that we don't need before passing data to the views. We were a little
unsure about this approach, and whether the code will get messy if
we're always omitting every unused key, but we decided to stick with it
for now. This keeps the `ProgrammeListItem` type representative of what
we need for the view, no less, no more, and ensures the what gets
passed to the view fully matches this type

We could have written the `omit` functionality ourselves, but unless
we start to see performance issues on the server, it's best not to
waste developer time and code space reinventing basic helper methods
that are well defined in open source libraries

[1]: https://github.com/ministryofjustice/hmpps-approved-premises-ui/commit/c1ef608c6eb38486251b60c2e9175c901bcb1c63#diff-634ec1902ba20e2ee91b4806fe88435d5cae6cd75c5605c90b3604bfb58d70f8R9-R17