Remove appWithAllRoutes test helper
We tend to use unit tests for controllers. I’ve replaced the one place
it’s used in in the errorHandler test with a smaller, more focussed stub
application, which hopefully should prevent misdirection.

This commit is shamelessly copied from the Approved Premises team's
project
[here](https://github.com/ministryofjustice/hmpps-approved-premises-ui/commit/a1bd26b95bc67b77bb0e796bf371b3295b04994b),
as we're trying to align our codebase with theirs as much as possible.