Add integration spec for find
This tests that list items with the right programme titles exist,
based on mocked data

This includes some changes to the integration test setup - both
the scripts in `package.json` and the integration test
environment file. The latter is renamed `integration.env`, and
the scripts are renamed to make it clearer what this is.
`npm run start-feature` and `feature.env` were a little opaque.
`ACCREDITED_PROGRAMMES_API_URL` is set in `integration.env` so
that when running integration tests, calls to the API are mocked

We're conditionally setting the Wiremock endpoint port to 9091
when running integration tests and 9092 when in development. The
latter will be used when we want to develop frontend behaviour
based on planned but incomplete/unmerged API behaviour

Co-authored-by: Ynda <yndajas@gmail.com>