Add djLint for linting Nunjucks
This adds the `djlint` package as an optional dependency. The Circle CI
build fails when including it as a dev dependency as the package has a
Python dependency, and the CI can't find the `pip` module to install
the djLint Python package

As a result, `package.json` now has both a `lint` script and a
`lint:ci` script. The former is used by `script/test` for checking
everything locally (including linting and formatting of Nunjucks files),
while the latter is used in the CI, only running eslint. Code could
therefore make it into production without passing djLint checks, but it
would be flagged locally when running the full test suite. Since these
checks are generally not about site-breaking issues, this seems
acceptable for now

This also and sets up some sensible defaults, such as allowing blank
lines in Nunjucks templates to avoid walls of text. There doesn't
appear to be a way to configure a particularly sensible approach to
this, so it's set to liberally allow all empty lines, even when
adjacent. There is the `blank_line_after_tag` option which enforces a
line after particular Nunjucks tags, but it doesn't do the same within
embedded HTML

I tried restricting the `lint` script's use of djLint to the
`server/views` directory by:
- in `.djlintrc`, specifying this directory in the `files` key
- in the `lint` script, using `-` instead of `.`. `-` references the
  path specified in `djlintrc`

However, this led to some unwanted behaviour. If formatting within
VSCode (e.g. on save), it would update the contents of the file to
"No files to check! 😢". It seems that setting affects VSCode
formatting too. So I've reverted just to specifying `.njk` files, but
in any directory within the repo. Performance seems fine in any case