Add djLint for linting Nunjucks
This adds the `djlint` package as an optional dependency. The Circle CI
build fails when including it as a dev dependency as the package has a
Python dependency, and the CI can't find the `pip` module to install
the djLint Python package

As a result, `package.json` now has both a `lint` script and a
`lint:ci` script. The former is used by `script/test` for checking
everything locally (including Nunjucks linter compliance), while the
latter is used in the CI, only running eslint. Code could therefore
make it into production without passing djLint checks, but errors would
be flagged locally when running the full test suite. Since these
checks are generally not about site-breaking issues, this seems
acceptable for now

Nunjucks automated formatting (in VSCode) was added in an earlier
commit, and also wouldn't be required before merging code

Since the formatter removes whitespaces around variables
(`{{variable}}`), the rule that specifies variables should be wrapped in
whitespace (`{{ variable }}`) is disabled