Enable djLint error highlighting in VSCode
When deciding to use Nunjucks Template Formatter for formatting and
djLint only for linting, the `files` key was added to `.djlintrc` to
be more specific about where to lint. When evaluating djLint as a
formatter (as well as a linter), this key was omitted as it would
produce unwanted behaviour in VSCode when using on-save autoformatting

However, it appears that with this key, VSCode also fails to highlight
linting errors in the editor. Removing the key fixes the behaviour,
allowing you to see errors without needing to run `npm run lint` or
`script test`. This therefore removes the key to enable Nunjucks linting
error highlighting