Sort and group import lines
43bdefc added sorting of members within multiple member import
declarations (e.g. `import {a, b, c} from 'alphabet'`), but not sorting
between lines. This adds that, as well as grouping external and internal
imports using a line break

It's possible to sort `import type` last, but then you lose the
external/internal divide. I've kept the latter for now, but we may want
to revisit this decision later