Allow skipping updates before running tests
Updating images and dependencies is a useful default to ensure we
aren't running tests with an old setup. However, when we know we're
up-to-date (or when it doesn't matter, e.g. if doing some TDD work),
it's useful to be able to skip this, which currently shaves 12s off
the runtime of `script/test` on my system