Add Prison and PrisonAddress types and factories
These represent the shape of a Prison we get back from the Prison API.
We'll be using this mostly to display the address details of a Prison on
a Course's Offering.

The factories here are a little confusing, as we've tried to ensure the
details on the top level Prison are carried through to the primary
`PrisonAddress` object, so we've had to use a few `or` statements here
to ensure we're passing the correct strings, rather than relying on
Fishery's ability to use params passed in over the defaults.