Run end-to-end tests on CI
This will run end-to-end tests on the deployed dev environment to ensure
the "real" service as a whole is working as expected, as opposed to our
Cypress integration tests which are running against a mock backend.

The script clones the E2E test repo to the CI environment, installs
dependencies and runs the Playwright test script.

Log in secrets are stored in the Circle CI env vars.