Add guard against user possibly being undefined
Here we add the `RequestWithUser` union type with the `AssertHasUser`
guard to be used in controller actions to assert that an authenticated
user will always be present in the request to each controller action.
This convinces the compiler that `request.user.token` is always going to
be available. In real life, if it's not present, the user will not be
authenticated, which HMPPS Auth will handle for us and prompt the user
to log in.

We've taken our approach from the accepted StackOverflow answer [here](
https://stackoverflow.com/questions/66614337/typescript-req-user-is-possibly-undefined-express-and-passport-js/66614362#66614362)

Co-authored-by: George <george@dxw.com>