Switch to Prison Register API
Auth

We had an issue with requests being unauthorised, which we solved by
pointing our local version of the Prison Register API at our local
HMPPS Auth instance by specifying the `OAUTH_ENDPOINT_URL`. We found
this by looking at the [HMPPS Registers UI repo][1]

Types

This also adds the `NonEmptyArray` type, which takes a type and
requires the array to have at least one element

The prison types previously each had a dedicated file, but as we won't
be creating too many types, they're combined into a single index file

The `OrganisationAddress and `PrisonAddress` types are also changed to
reflect the fields that can be `null`, based on the current Prison
Register dev data. We may want to update this to work with any fields
being `null`, in case such data is added in future. The fields are
always present even if `null`, so the `?` is removed from
`addressLine2`

[1]: https://github.com/ministryofjustice/hmpps-registers/blob/0e15ad772bfb521701899e886e4eac180bb6121c/docker-compose.yml

Co-authored-by: George <george@dxw.com>