Get docker-compose working locally with new DB
This started failing once we added the database to the API, as we
weren't starting a postgres database.

I wanted to add a condition to wait for `hmpps-auth` to be healthy
before starting the server, but it appears to timeout after about 90
seconds (it takes a long time to start up!) so I've taken this out.