Style dashboard
This is based on the dashboard pattern used in [HMPPS Auth][1] and a few
other MoJ services. The styling is simplified a little, removing
redundant declarations and renaming classes both to suit the domain and
provide greater clarity

The pattern on HMPPS Auth uses JavaScript to make it so that clicking
anywhere on the card will send a click event to the anchor tag. We do the
same here, but we make the JavaScript-disabled styling clearer, leaving
`cursor: pointer`, `:active { bottom: -5px }` and other styles off the
card unless a class is added by JavaScript

[1]: https://github.com/ministryofjustice/hmpps-auth/blob/c3930cd0a415a3cee52b4c26116587964702db5d/src/main/resources/static/css/app.css#L50