Fix prerequisite display order
This fixes the order in which course prerequisites are displayed, per
the designs, and also updates the prerequisite names to match the current
data

The course factory defaults to ordering the prerequisites
alphabetically, which is different to the order we want them presented
in. Ultimately, the order we get them in from the API shouldn't matter:
we should display them consistently regardless of the order we get them
in

The `courseUtils` spec purposefully duplicates the alphabetical order of
prerequisites that is the factory default in order to demonstrate that
the order of `prerequisiteSummaryListRows` should not be determined by
the order of the prerequisites themselves, but rather a set order as
defined in the `prerequisiteSummaryListRows` method. This duplication
also protects the test from changes to the factory defaults

The integration tests still work as they're testing that the page has
the summary list items in the order of the prerequisite summary list
items on the course presenter, rather than in the order of the
unmanipulated prerequisites on the course