Don't run express app with docker compose
The `app` container (the express app) in the `docker-compose.yml` would
be run alongside all of the other containers specified in the config
when running `docker compose up`. However, in our
`script/utils/start_backing_services` script, we exclude the app itself
when starting the backing services with (`--scale=app=0`) and run it
with `npm run start:dev` in `script/server`, so the dockerised version
is never being used.

The dockerised version also isn't currently working when running all
containers without the `SENTRY_DSN` environment variable passed to it.
Rather than adding this variable, it feels more appropriate to remove it
as we're never using it to start the server locally.

We can always add it back in if we do decide we want to run the whole
project via Docker Compose, but that seems unlikely in the near future.