Add `initialiseTitle` util
`initialiseName` already exists in `utils` but this takes a name, e.g.
`Karl Marx` and initialises just the first name, returning `K. Marx` in
this example

`initialiseTitle` meanwhile initialises all words, with space as the
delimiter, and with capitalisation on each, so e.g.
`The Teachings of Karl Marx` would become `TTOKM`

This will be used in the course factory to generate the
`alternativeName` from the `name`