Use index for utils
Compared to other index changes, this one is a little more substanial.
Previously we were importing specific methods from specific utils,
whereas if we want to use an index, we have a few options:
- import and export each method in the index, but when importing from
  the index their context/informal namespacing will be lost
- export a single object from each util containing all its public
  methods, which is then made accessible by the index. This makes the
  namespacing even more prominent than before, by showing it on every
  method call, not just on import lines. It also means importing the
  entire util even when only one or two methods may be used

The latter is the approach taken here. Given the utils are fairly
lightweight, importing the full util is likely not too costly. It's
also possible that in compilation, the unused methods/redundancy will
be stripped

One outstanding issue is that `azureAppInsights`, which is in the
`utils` folder, doesn't play nice with this approach: the find
integration tests fail when trying to apply it to this file. For now,
this is the only file that continues to use the old non-index approach