Use index for middleware
This is used in `server/app.ts`, but not in
`server/utils/routeUtils.ts`, which is the only other place that imports
from middleware. In that file, `asyncMiddleware` is imported directly in
order to avoid a cyclical dependency lint error. Hopefully that issue is
isolated to this specific context and doesn't reoccur with the shift to
using indices by default