Fallback on page heading as page title
Where possible, we should use the same text for the page heading as the
page title

This replaces the `pageTitle` variable with a `customePageTitleEnd`
variable, which can be set in views like the former. If provided, it
will be appended to "HMPPS Accredited Programmes" (with a dash
separator). If not provided, the page heading will be appended instead.
If there's no custom page title end or page heading, nothing will be
appended

The integration test suite's page class is adapted to this new approach,
testing the page title regardless of how it's constructed, provided
we're on a service-internal page (e.g. not HMPPS Auth)