Concatenate prerequisites of the same type
A single course can have multiple prerequisites of the same type (name).
Previously we were looping over the prerequisites and inserting them
into an array of summary list rows according to a predefined order. If
there were duplicates, only the last in the array of that type would be
retained (since each prerequisite would overwrite any previous of its
type)

This adds conditional logic so that concatenation occurs should a
prerequisite of the same type have already been added to the summary
list rows array