Display secondary offering email if one exists
We've now added a secondary offering email to the API for Programmes
that might have more than one contact email address.

If one is present, we render the item in the summary list on the
Offering page, otherwise we don't render the summary list item at all
(as we don't want to display an empty item).

There's a fair bit going on in this commit as I've renamed some
functions and changed return values of types to support a
possibly-undefined summary list item, but I wanted to make sure all
tests were passing.

The integration tests here aren't exactly DRY, but we can refactor them
once we have a bit more time if we think it's necessary.