Join multiple prison categories to single string
Although prison categories come back from the Prison Register as an
array, we want to represent them on our Organisation model with a single
string.

I debated renaming `category` to `categories` on our internal
Organisation model here, but I opted to leave it as category as I
believe semantically you could refer to a an organisation as a "Category
A and B prison" and aligns with the table header in our designs.