Fix E2E CI artifacts path
When running tests in the E2E repo, two directories are created:
`playwright-report` and `test-results` (not `test_results`). The former
includes the HTML report, while the latter is empty. Approved Premises
[recently fixed this][1]

The `destination` defaults to the `path` (see [here][2]), so there's no
need to specify this when they're the same - we don't do this elsewhere

[1]: https://github.com/ministryofjustice/hmpps-approved-premises-ui/commit/7aee650ae4d24f97dc4e88587944cddde7180994
[2]: https://circleci.com/docs/artifacts/#uploading-artifacts