Add cypress-axe package
... along with axe-core, and configure for TypeScript

We'll use cypress-axe to run automated accessibility testing with our
integration test suite

The [Regulated Professions Register PR][1] that added cypress-axe didn't
add axe-core, but the cypress-axe documentation suggests it's needed,
and we'll use it for typing later. Approved Premises added it when adding
cypress-axe

After adding `cypress-axe` to the integration tests' `tsconfig.json`,
there was a TypeScript error with `typeRoots` like [this one][2]. The
fix used there didn't work in this case, but it seems we don't need
`typeRoots`. It's not entirely clear why we are using them (the relevant
template commit isn't verbose), but it may be to include everything from
`./server/types` - we include specific types through the aliases we set
up in `paths`. Of the `typeRoots` that are removed here, the first was
added in the initial commit and the second and third were added to [fix
a similar error][2] that doesn't reappear if we remove all `typeRoots`.
All tests pass without them. [Very few MOJ projects use them][4]

[1]: https://github.com/UKGovernmentBEIS/regulated-professions-register/pull/145
[2]: https://github.com/ministryofjustice/hmpps-accredited-programmes-ui/pull/69/commits/7907980378e530b4fdf19c1675866534de14ec3c
[3]: https://github.com/ministryofjustice/hmpps-accredited-programmes-ui/commit/7907980378e530b4fdf19c1675866534de14ec3c
[4]: https://github.com/search?q=org:ministryofjustice++typeRoots+path:integration_tests/tsconfig.json&type=code