Log accessibility violations
The current setup allows us to fail integration tests where there are
accessibility (WCAG) violations using the `checkAccessibility` command,
but it doesn't seem to tell us anything about the violation. Using the
UI, you can see which element is in violation of a rule, but not why.
The automated tests in axe DevTools seem not always to report a
violation on the same element

This adds logging to make it clear which rules have been violated
(though it probably doesn't tell you where, so you might need to combine
it with the UI-based testing)

This is mostly adapted from the cypress-axe GitHub README (['Using the
violationCallback argument'][1] and ['In your spec file'][2]), but also
[another MoJ project][3]

[1]: https://github.com/component-driven/cypress-axe#using-the-violationcallback-argument
[2]: https://github.com/component-driven/cypress-axe#in-your-spec-file
[3]: https://github.com/ministryofjustice/bichard7-next-ui/commit/db4e33e545a64fe5f4c331191de8638a97f9e1fa