Exclude phase banner from axe tests
Per [this Regulated Professions Register commit][1] this fails axe tests
because of the rule "All page content must be contained by landmarks"

It's unclear how the rules that were violated were identified in that
PR, since it doesn't appear to add any logging of these. Here we've
added the logging, and when running the tests in the Cypress UI, we can
see that clicking on the error highlights the phase banner

Again according to that PR, the violation is due to the banner being
outside `<header>` and `<main>` elements. Apparently this is a [known
issue][2] with the component, but as the region rule is not a
requirement for WCAG, it should be okay to ignore for now

The approach here differs to that PR - it uses `cy.configureAxe`, which
feels like a good place to put config, rather than passing a
selector-based exception to `cy.checkA11y`. The exception is more
specific: a selector and a specific rule, so that other rules are still
checked. I'm not sure if this is possibly with `cy.checkA11y`

[1]: https://github.com/UKGovernmentBEIS/regulated-professions-register/commit/33be548361c21ff9dfd04c65d7a1b6374230fcdc
[2]: https://github.com/alphagov/govuk-frontend/issues/1604

use configure instead