Disable no empty function rule for constructors
This rule is enabled by default with the latest version of
typescript-eslint.

I've disabled it for constructors in general as it seems to complain
when using property parameters, for which I think an empty function is
fine.

I've updated the one place we had an empty function in the service where
this rule was failing to add a comment as that appears to be the "done
thing" in TypeScript according to [the documentation][0].

[0]: https://eslint.org/docs/latest/rules/no-empty-function#rule-details