Expose find paths in Nunjucks
This loosely follows a pattern in Approved Premises, though there the
equivalent would be:

```
  njkEnv.addGlobal('paths', {
    ...findPaths,
  })
```

I've added a `find` key at the top level to avoid any future naming
clashes, e.g. if `findPaths` and `referPaths` each had a `courses` key,
to avoid one overwriting the other. This kind of scenario might be
unlikely, but it feels like a good guard to have

The first use for this `paths` global will be to provide `href`s for
back links without needing the controller to pass down more information