Add back links
These are added to a new block called `backLink`. Per [this commit][1]
in Approved Premises and the guidance linked in the [related PR
description][2], back links should come before `main` to allow 'Skip to
main content' links to skip them (although we don't currently have any
of these links). Apparently this can lead to an axe violation, but I'm
not seeing it here locally, so I haven't disabled anything for now

Here we use the block `backLink` instead of just `beforeContent` in
order to retain elements added to `beforeContent` in the layout (the
phase banner). `backLink` is placed at the bottom of `beforeContent` in
the layout

[1]: https://github.com/ministryofjustice/hmpps-approved-premises-ui/pull/797/commits/a9acb31d446fc204951ab6aa4ae6b44353752ffb
[2]: https://github.com/ministryofjustice/hmpps-approved-premises-ui/pull/797