Modify CSP to allow forms to target HMPPS Auth
Update the Content Security Policy to allow the HMPPS Auth URL as a
possible form action target.

Currently, if a 403 error occurs on a GET request, this will be captured
by the error handling setup in errorHandler.ts, and the user will be
redirected to the sign out URL, which then redirects to the HMPPS Auth
URL. However, if a 403 error occurs on a POST request, this second
redirect may not occur, and the user may, depending on their choice of
browser, be frozen on the form page they just submitted.

This was encountered when trying to submit a form without the `csrf`
token as an hidden `input` but could occur if the token has expired.

Due to CSP implementation details that vary between browsers, adding the
HMPPS Auth URL to our form action targets allows this second redirect to
occur as expected.