Display flash message after setting cookies
This uses `connect-flash` to display a confirmation banner when cookie
preferences have been saved.

This sets a `returnTo` url on the request object to allow users to
redirect back to the page they were on, which is also used to provide
the URL for effectively reloading the page to hide the flash message.

It's slightly hacky in that we only set this `returnTo` session variable
when not posting to the `cookiePreferencesPostRoute` as otherwise this
would be set to the `/cookie-preferences` post route when the request is
intercepted by our cookie middleware, which prevents us being able to
save the correct URL, but I think this will have to do for now.