Simplify data exports
The type `DataAccess` wasn't used anywhere, so this is removed

`dataAccess` returned a function that returns an object - instead here
the object's key/value pairs are exported directly, like everything else

Since `hmppsAuthClient` was previously inside a function, this is now
changed to be a builder function like the other clients to retain
similar on-demand instantiation. However, previously I believe it would
just have been generated once or twice when `services()` is called in
`server/index.ts`, whereas now it's generated every time the builder
is called like other clients. All tests pass with this change, so this
seems okay