Use index for utils
Compared to other index changes, this one is a little more substantial.
Previously we were importing specific methods from specific utils. If we
want to use an index, we have a few options:
- import and export each method in the index, but when importing from
  the index their context/informal namespacing will be lost
- export a single object from each util containing all its public
  methods, which is then made accessible by the index. This makes the
  namespacing even more prominent than before, by showing it on every
  method call, not just on import lines. It also means importing the
  entire util collection (e.g. `organisationUtils`) even when only one
  or two methods may be used

The latter is the approach taken here. Given the utils are fairly
lightweight, importing a full util collection is likely not too costly.
It's also possible that in compilation, the unused methods/redundancy
will be stripped

Compared with an earlier version of this commit (7ee403b), this makes
some additional changes to `typeUtils` and a change to `@types/express`
to fix a new TypeScript error:

```
server/utils/typeUtils.ts:1:71 - error TS2694: Namespace 'global.Express' has no exported member 'RequestWithUser'.

1 const assertHasUser: (req: Express.Request) => asserts req is Express.RequestWithUser = req => {
                                                                        ~~~~~~~~~~~~~~~
```

One outstanding issue is that `azureAppInsights`, which is in the
`utils` folder, doesn't play nice with this approach: the find
integration tests fail when trying to apply it to this file. For now,
this is the only file that continues to use the old non-index approach