Add referral start page
We will add feature flag integration tests in a separate branch later,
since it will require quite a bit of setup

We've gone for a flatter path structure here than we did with find. Find
paths will be updated in a later commit. All refer pages are about referrals
and all find pages are about courses, so this nesting is unnecessary