Add Prisoner client
with a `getPrisoner` method which currently searches the Prisoner
Offender Search for a Person in Prison by their Prison number. In
future, we'll want to expand this to support searching by name and date
of birth but this is a first step.

According to the DPS style guide[1], we should be using the term "Prison
Number" rather than "Prisoner Number". In later commits, I've used
Prison Number internally on our `Person` model, but kept the term
"Prisoner Number" when referring to the key returned from the Prisoner
Offender Search, as this is how it is referred to in that service. This
may potentially be a bit confusing to future readers, but it feels right
to use the preferred term in our own service.

This also adds a `Prisoner` type and factory to generate this for use in
tests.

I've used the term Prisoner here to represent the object returned by the
API, even though MOJ would prefer to move away from this label towards
"Person in Prison". We'll use "Person in Prison" for any user-facing
content.

[1]:
https://sites.google.com/digital.justice.gov.uk/dps-content-style-guide?pli=1&authuser=1#h.sp9fg4lfnk2k