Add navigation
This adds a primary navigation component with a link to the list of
programmes. It seems like this typically goes above the phase
banner/directly below the header in most HMPPS/MoJ designs, but here we
want it to go below. It's therefore in the `beforeContent` block, and
needs similar margin/padding adjustments to the dashboard page (and a
little extra) to be full width