Add PeopleController and /confirm-person template
This also adds a new function to `personUtils` to format things into the
summary list shape for display in the UI.

I toyed with the idea of creating a whole `PersonPresenter` type that
would encapsulate the title and summary list functionality here as we
have with a `CoursePresenter`, but as we currently only need to display
a single summary list, I thought it was over-engineering it a bit. We
can always refactor this to a more fully-fledged Presenter if we need to
add additional functionality in the future.

And hook the new controller up with routes for the new section of the
refer journey.