Add PersonUtils with personFromPrisoner function
This converts a Prisoner object we received from the Prisoner Offender
Search into a Person object for our own internal representation,
converting the necessary fields.

There are some fields that may be missing. Hopefully I've covered all of
them here, but it's hard to tell from the data in Dev.

I've opted to display: "Not entered" where information is missing, as
this is what is displayed in DPS in this case.