Allow custom test in git history checking
The check for an exit code of 1 is now changed to non-zero. I found that
`tsc` (typecheck in TypeScript) exits with a code of 2 if there are errors.
It feels safer to assume test suites will exit with 0 if there's no issue,
rather than that the exit code will always be 1 if there is