Mock Prisoner Offender Search locally
We decided that getting the Prisoner Offender Search to run locally as a
backing service would require too much overhead due to the complex
dependencies we'd also need to run locally.

Because of this, I've opted to mock the endpoints we use when running
locally alongside other backing services.

We can also configure our local environment to use the `dev` environment
for the Prisoner Offender Search, but this requires developers to be
connected to the MOJ VPN to access the service. This provides the
ability to develop locally without needing to connect to the VPN.