Configure ESLint Member Ordering
This currently only applies to classes, but will sort methods in the
order defined in the `memberTypes` key.

We've decided to turn this on as there can be a fair bit of overhead
on PR review, and catching this kind of thing early can save us time
in the long run.

This fixes the alphabetical order of all groups in the offending
classes.

A later commit will refactor files like utils into classes, at which
point this rule will apply.

Co-authored-by: Ynda Jas <yndajas@gmail.com>