Add template referral form page
This is a placeholder page with the course name, prison location and
strand that we'll be building on in the next few PRs.

It's currently available at
`/offerings/:offeringId/referrals/:referralId`, although the
`referralId` isn't used yet. This is slightly different from previous
URLs in the journey that use `refer` as their action, as now that the
referral has been created, `/referrals/` feels like more sensible
resource name to be reading, rather than `refer/:id`.

I've kept it as simple as possible to allow us to get it up and running
quickly.