Combine "Find person" integration tests
This gives us a more complete user journey when testing the application.

We're only testing the happy path here, as we've not yet got any custom
behaviour for showing error messages if a user isn't found, but we'll
add test cases to cover that scenario when we add those error messages
in the near future.