Check for person details in integration test
The task list page now has a person details banner. This checks that it
appears. Some extra stubbing is needed here since it's the first time
we're asserting anything about the person attached to a referral

I've added this `should` method to `Page` rather than `TaskListPage`
since the person details banner will be used throughout most of the
Refer journey