Add referral task list util method
This returns task list items in the format needed for an MOJ task list.
Currently the referral model doesn't have the fields on it to make this
a particularly dynamic method, so the return value is mostly hardcoded,
but it will act as a template that we can update once we have the
required model fields

We're using the `moj-task-list__task-completed` class on tags here even
when the status is not completed. This class actually just floats the
tags to the right of the task list row and adds margin. We could
recreate this class in our own stylesheets and give it a more accurate
name, but I've favoured using a shared class here. The same is done in
[the Interventions UI][1]

There's some logic here, similar to Interventions, to allow task list
items not to have links. Currently we don't want the "Confirm personal
details" task to have a link, since it will be completed en route to
this page. A link may be added later to reshow the personal details
(but not allow updates)

Explanation of why we're building this manually instead of using the
[Nunjucks macro][2] will come in the next commit

[1]: https://github.com/ministryofjustice/hmpps-interventions-ui/pull/34/files#diff-03f5907738cb33527fe8f34fbe0d5b1b7bb6cef5bb8cdec11da3d40731b28ba2R7
[2]: https://design-patterns.service.justice.gov.uk/patterns/task-list