Add update referral client method
This requires a new type which includes all (and only) the updatable
fields on a Referral. This endpoint doesn't return an object, so the
client method and tests are a little different. I'm following what
Approved Premises do with various client methods that have body-less API
responses here (though `update` isn't one of them)