Add applicationSummaryListRows method to ReferralUtils
This will be used in the check your answers page to display a summary of the application to the referrer before submission

Using the current users username for the Referrer name for now until we identify where to get their full name from.  Also ommiting Contact number which is shown in the designs as unsure if we have this information at the moment.