Update Cypress config for v13
`videoUploadOnPasses` has been removed - we already delete videos where there
are no failures using an `after:spec` Node event, largely following an example
in the Cypress docs, so this was likely redundant

`video: true` - `video` now defaults to `false`, and we do want to continue
recording video (and retaining it for failures)

`videoCompession: true` - `videoCompression` now defaults to `false`. Setting it
to true returns it to compression at a constant rate factor of 32 (you can also
provide a specific number between 1 and 51). I don't really know what this
means, but with compression a video was 421 kB vs 2.7 MB without, and the
compression run was marginally quicker (based on a sample of one run per
condition)