Add `MiddlewareUtils` class
This currently contains a single function for decoding the roles on an
HMPPS Auth token.

This behaviour is moved across from the `AuthMiddleware` function as
we'll be using it to add roles to the `User` object in the service to
determine what they have access to do.

I thought about including this in the `utils` directory but decided to
package it with the middleware-focused classes to keep it separate and
make it obvious that these utils are only to be used in middleware. If
we need to use them elsewhere, we can move them in future.