Fix Cypress config TypeScript error
In 89396d7, a `mockedUsername` abstraction was added to make it clear what
`USER1` is when used in various places

This was added to the file's default export, which is spread as part of the
object in `on('task', [object])` within `cypress.config.ts`. When opening
the config file in VS Code, a TypeScript error is displayed (we don't
typecheck this file, so it's easily missed). The error indicates that the
`mockerUsername` string value doesn't typecheck. Changing `mockedUsername` to
a function - like most if not all of the other spread values - fixes the
error