Fix local/integration Docker conflict
Before this commit, if you had the local server running then ran the
integration tests (I think via `npm` directly or via `script/test`), the
local server would stop working and would need a restart

Debugging this was a little complicated

**Issue 1:** removing orphans removes all other running containers
**Solution:** don't `--remove-orphans` in `script/test`

**Issue 2:** we rely on Wiremock locally for the Refer journey, but the
integration tests replace the Wiremock container and reset the stubs
**Solution:** use a separate Docker project for integration tests, and
within that project a unique Wiremock container

We can use a separate Docker project by passing in a name to
`docker-compose` with `-p`. When doing this, we need to use unique
container names and ports to avoid cross-project conflict. The container
names are therefore appended with `_integration-test`, the same as the
project

We couldn't simply use unique container names/ports and the same
project, since Docker will only run one instance of a given image per
project - any time you ask for a new instance to be created, even with a
different container name, it will replace the old instance

We don't currently tear down the containers if running
`npm run test:integration:*` - we might want to add a
`close-backing-services` script to complement the
`start-backing-services` one to avoid container hanging around

The CI still uses the same Redis port as before, so a separate `package.json`
script is added to customise the environment variables when running the
integration test server there